Fixed issue with the webpack grunt task where it was not reporting the errors correctly
